+++ .github/workflows/pr-checks.yml
@@ -32,8 +32,6 @@ jobs:
         version: 9
     - name: Create a link to envtest in operator/bin
       run: mkdir -p ./operator/bin && ln -s $(which setup-envtest) ./operator/bin/setup-envtest
-    - name: Install goyacc
-      run: go install golang.org/x/tools/cmd/goyacc@latest
     - name: Create the operator manifests
       run: make operator-manifests
     - name: Run tests
